requirements.txt
streamlit
pandas
numpy
matplotlib
seaborn
scikit-learn
mysql-connector-python
bcrypt
